requirements.txt
Django==1.11
djangorestframework==3.6.3
docker
django-cors-headers